+++ pkg/cps/dv_challenges.go
@@ -29,30 +29,30 @@ type (
 
 	// DV is a Domain Validation entity
 	DV struct {
-		Challenges         []Challenges `json:"challenges"`
-		Domain             string       `json:"domain"`
-		Error              string       `json:"error"`
-		Expires            string       `json:"expires"`
-		RequestTimestamp   string       `json:"requestTimestamp"`
-		Status             string       `json:"status"`
-		ValidatedTimestamp string       `json:"validatedTimestamp"`
-		ValidationStatus   string       `json:"validationStatus"`
+		Challenges         []Challenge `json:"challenges"`
+		Domain             string      `json:"domain"`
+		Error              string      `json:"error"`
+		Expires            string      `json:"expires"`
+		RequestTimestamp   string      `json:"requestTimestamp"`
+		Status             string      `json:"status"`
+		ValidatedTimestamp string      `json:"validatedTimestamp"`
+		ValidationStatus   string      `json:"validationStatus"`
 	}
 
-	// Challenges contains domain information of a specific domain to be validated
-	Challenges struct {
-		Error             string              `json:"error"`
-		FullPath          string              `json:"fullPath"`
-		RedirectFullPath  string              `json:"redirectFullPath"`
-		ResponseBody      string              `json:"responseBody"`
-		Status            string              `json:"status"`
-		Token             string              `json:"token"`
-		Type              string              `json:"type"`
-		ValidationRecords []ValidationRecords `json:"validationRecords"`
+	// Challenge contains domain information of a specific domain to be validated
+	Challenge struct {
+		Error             string             `json:"error"`
+		FullPath          string             `json:"fullPath"`
+		RedirectFullPath  string             `json:"redirectFullPath"`
+		ResponseBody      string             `json:"responseBody"`
+		Status            string             `json:"status"`
+		Token             string             `json:"token"`
+		Type              string             `json:"type"`
+		ValidationRecords []ValidationRecord `json:"validationRecords"`
 	}
 
-	// ValidationRecords represents validation attempt
-	ValidationRecords struct {
+	// ValidationRecord represents validation attempt
+	ValidationRecord struct {
 		Authorities []string `json:"authorities"`
 		Hostname    string   `json:"hostname"`
 		Port        string   `json:"port"`